fastapi-backend/requirements.txt
fastapi[all]
aiofiles
aiohttp
pillow
redis
pymilvus
insightface
tqdm
opencv-contrib-python
opencv-python
opencv-python-headless
#onnxruntime-gpu
#onnxruntime
numpy
matplotlib
moviepy
scikit-learn
python-dotenv
pyjwt
passlib
pymongo